V3.0.0_20241120230136

+++ include/net.h
@@ -551,7 +551,7 @@ extern char	net_boot_file_name[1024];/* Boot File name */
 /* Indicates whether the file name was specified on the command line */
 extern bool	net_boot_file_name_explicit;
 /* The actual transferred size of the bootfile (in bytes) */
-extern u32	net_boot_file_size;
+extern long	net_boot_file_size;
 /* Boot file size in blocks as reported by the DHCP server */
 extern u32	net_boot_file_expected_size_in_blocks;
 
@@ -587,7 +587,7 @@ extern int net_ntp_time_offset;			/* offset time from UTC */
 
 /* Initialize the network adapter */
 void net_init(void);
-int net_loop(enum proto_t);
+long net_loop(enum proto_t);
 
 /* Load failed.	 Start again. */
 int net_start_again(void);
